+++ rally_openstack/common/cfg/opts.py
@@ -20,7 +20,6 @@ from rally_openstack.common.cfg import ironic
 from rally_openstack.common.cfg import magnum
 from rally_openstack.common.cfg import manila
 from rally_openstack.common.cfg import mistral
-from rally_openstack.common.cfg import monasca
 from rally_openstack.common.cfg import neutron
 from rally_openstack.common.cfg import nova
 from rally_openstack.common.cfg import octavia
@@ -43,7 +42,7 @@ def list_opts():
 
     opts = {}
     for l_opts in (cinder.OPTS, heat.OPTS, ironic.OPTS, magnum.OPTS,
-                   manila.OPTS, mistral.OPTS, monasca.OPTS,
+                   manila.OPTS, mistral.OPTS,
                    nova.OPTS, osclients.OPTS, profiler.OPTS,
                    vm.OPTS, glance.OPTS, watcher.OPTS, tempest.OPTS,
                    keystone_roles.OPTS, keystone_users.OPTS, cleanup.OPTS,