+++ rally_openstack/common/consts.py
@@ -55,7 +55,6 @@ class _Service(utils.ImmutableMixin, utils.EnumMixin):
     NEUTRON = "neutron"
     DESIGNATE = "designate"
     CEILOMETER = "ceilometer"
-    MONASCA = "monasca"
     S3 = "s3"
     TROVE = "trove"
     SWIFT = "swift"
@@ -85,7 +84,6 @@ class _ServiceType(utils.ImmutableMixin, utils.EnumMixin):
     NETWORK = "network"
     DNS = "dns"
     METERING = "metering"
-    MONITORING = "monitoring"
     S3 = "s3"
     DATABASE = "database"
     OBJECT_STORE = "object-store"
@@ -115,7 +113,6 @@ class _ServiceType(utils.ImmutableMixin, utils.EnumMixin):
             self.NETWORK: _Service.NEUTRON,
             self.DNS: _Service.DESIGNATE,
             self.METERING: _Service.CEILOMETER,
-            self.MONITORING: _Service.MONASCA,
             self.S3: _Service.S3,
             self.DATABASE: _Service.TROVE,
             self.OBJECT_STORE: _Service.SWIFT,